requirements.txt
anndata>=0.9
scanpy>=1.9
scvi-tools>=1.0
xgboost>=1.7
scikit-learn>=1.2
numpy>=1.23
pandas>=1.5
scipy>=1.9
joblib>=1.2
matplotlib>=3.6
seaborn>=0.12
# optional
optuna>=3.0
